# Swagger-OpenAPI/dist/openapi-bank.yaml
openapi: 3.0.3
info:
  title: REST-API-BANK
  description: Full stack web project
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Dev server
     
paths:
  /customers/login:
    post:
      summary: Customer login
      tags:
        - Customers
      operationId: loginCustomer
      requestBody:
        description: Customer credentials for authentication
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'       

  /customers/register:
    post:
      summary: Customer registration
      tags:
        - Customers
      operationId: registerCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustRequest'
      responses:
        '201':
          description: Customer successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustResponse'
        '409':
          description: Customer already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error' 

  /transactions:
    post:
      summary: Create transaction
      tags:
        - Transactions
      operationId: createTransaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
              properties:
                amount:
                  type: number
                  example: 150.50
      responses:
        '200':
          description: Transaction successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionsResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # /transactions:
    get:
      summary: Get transactions by filter
      tags:
        - Transactions
      operationId: getTransactionsByFilter
      security:
        - bearerAuth: []
      parameters:
        - name: amount
          in: query
          required: false
          schema:
            type: number
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Transactions list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionsResponse'
        '400':
          description: Invalid filter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /transactions/{id}:
    get:
      summary: Get transaction by ID
      tags:
        - Transactions
      operationId: getTransactionById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaction found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionsResponse'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Update transaction
      tags:
        - Transactions
      operationId: updateTransaction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
              properties:
                amount:
                  type: number
                  example: 200
      responses:
        '200':
          description: Transaction updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionsResponse'
        '400':
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete transaction
      tags:
        - Transactions
      operationId: deleteTransaction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaction deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionsResponse'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # /transactions:
  #   get:
  #     summary: Get transactions by filter
  #     tags:
  #       - Transactions
  #     operationId: getTransactionsByFilter
  #     security:
  #       - bearerAuth: []
  #     parameters:
  #       - name: amount
  #         in: query
  #         required: false
  #         schema:
  #           type: number
  #       - name: date
  #         in: query
  #         required: false
  #         schema:
  #           type: string
  #           format: date
  #     responses:
  #       '200':
  #         description: Transactions list
  #         content:
  #           application/json:
  #             schema:
  #               $ref: '#/components/schemas/TransactionsResponse'
  #       '400':
  #         description: Invalid filter
  #         content:
  #           application/json:
  #             schema:
  #               $ref: '#/components/schemas/Error'

components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    CustRequest:
      type: object
      required:
        - name
        - password
      properties:
        name:
          type: string
          minLength: 3
          example: Alex
        password:
          type: string
          minLength: 6
          example: 1234Abcd

    CustResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...

    Transaction:
      type: object
      properties:
        id:
          type: integer
          example: 1
        amount:
          type: number
          format: double
          example: 10.55
        date:
          type: string
          format: date-time
          example: 2024-01-01T12:00:00Z
        customerId:
          type: integer
          example: 1

    TransactionsResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'

    Error:
      type: object
      properties:
        status:
          type: integer
          example: 400
        message:
          type: string
          example: Invalid input
        timestamp:
          type: string
          format: date-time
          example: 2024-01-01T12:00:00Z